refactor(services): reorganize services into specific subpackages
Move BidList related services into subpackage services.bidlist.

Update imports in controllers, tests.
